migrate lambda S3 buckets to remote-state Terraform definition
This is a preliminary step toward the goal of separating the Lambda
build procedure from the Terraform infrastructure provisioning.
Presently, one cannot initially deploy the Terraform infrastructure
because the AWS S3 buckets containing the Lambda artifacts do not
exist yet. When the buckets are provisioned separately from the
rest of the infrastructure, the artifacts can be build an uploaded
prior running `terraform apply`.